+++ ptai-teamcity-plugin/ptai-teamcity-plugin-common/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/plugin/teamcity/Labels.java
@@ -54,7 +54,8 @@ public class Labels {
     public static final String FAIL_IF_FAILED = "Fail step if SAST failed";
     public static final String FAIL_IF_UNSTABLE = "Fail step if SAST unstable";
 
-    public static final String VERBOSE = "Verbose logging";
+    public static final String FULL_SCAN_MODE = Resources.i18n_ast_settings_fullScanMode_label();
+    public static final String VERBOSE = Resources.i18n_ast_settings_verbose_label();
     public static final String INCLUDES = "Files to analyse";
     public static final String REMOVE_PREFIX = "Remove prefix";
     public static final String EXCLUDES = "Exclude files";